+++ tests/Checkout/ExpressCheckout/ExpressCheckoutSubscriberTest.php
@@ -10,8 +10,6 @@ namespace Swag\PayPal\Test\Checkout\ExpressCheckout;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Checkout\Cart\LineItem\LineItem;
 use Shopware\Core\Checkout\Cart\SalesChannel\CartService;
-use Shopware\Core\Checkout\Payment\PaymentMethodCollection;
-use Shopware\Core\Checkout\Payment\PaymentMethodEntity;
 use Shopware\Core\Content\Cms\CmsPageCollection;
 use Shopware\Core\Content\Cms\CmsPageEntity;
 use Shopware\Core\Content\Cms\Events\CmsPageLoadedEvent;
@@ -45,8 +43,6 @@ use Swag\PayPal\Checkout\ExpressCheckout\ExpressCheckoutSubscriber;
 use Swag\PayPal\Checkout\ExpressCheckout\Service\PayPalExpressCheckoutDataService;
 use Swag\PayPal\Setting\SwagPayPalSettingStruct;
 use Swag\PayPal\Test\Helper\ServicesTrait;
-use Swag\PayPal\Test\Mock\Repositories\PaymentMethodRepoMock;
-use Swag\PayPal\Test\Mock\Repositories\SalesChannelRepoMock;
 use Swag\PayPal\Test\Mock\Setting\Service\SettingsServiceMock;
 use Swag\PayPal\Util\PaymentMethodUtil;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
@@ -193,6 +189,22 @@ class ExpressCheckoutSubscriberTest extends TestCase
         static::assertNull($actualExpressCheckoutButtonData);
     }
 
+    public function testAddExpressCheckoutDataToPageWithInactivePaymentMethod(): void
+    {
+        $salesChannelContext = $this->createSalesChannelContext(true, false);
+        $event = new CheckoutCartPageLoadedEvent(
+            new CheckoutCartPage(),
+            $salesChannelContext,
+            new Request()
+        );
+
+        $this->getExpressCheckoutSubscriber()->addExpressCheckoutDataToPage($event);
+
+        /** @var ExpressCheckoutButtonData|null $actualExpressCheckoutButtonData */
+        $actualExpressCheckoutButtonData = $event->getPage()->getExtension(ExpressCheckoutSubscriber::PAYPAL_EXPRESS_CHECKOUT_BUTTON_DATA_EXTENSION_ID);
+        static::assertNull($actualExpressCheckoutButtonData);
+    }
+
     public function testAddExpressCheckoutDataToPageWithInvalidSettings(): void
     {
         $salesChannelContext = $this->createSalesChannelContext();
@@ -276,6 +288,19 @@ class ExpressCheckoutSubscriberTest extends TestCase
         );
     }
 
+    public function testAddExpressCheckoutDataToCmsPageCmsWithInactivePaymentMethod(): void
+    {
+        $event = $this->createCmsPageLoadedEvent(true, false);
+
+        $this->getExpressCheckoutSubscriber()->addExpressCheckoutDataToCmsPage($event);
+
+        $cmsPage = $event->getResult()->first();
+        static::assertNotNull($cmsPage);
+        /** @var ExpressCheckoutButtonData|null $actualExpressCheckoutButtonData */
+        $actualExpressCheckoutButtonData = $cmsPage->getExtension(ExpressCheckoutSubscriber::PAYPAL_EXPRESS_CHECKOUT_BUTTON_DATA_EXTENSION_ID);
+        static::assertNull($actualExpressCheckoutButtonData);
+    }
+
     public function testAddExpressCheckoutDataToCmsPageCmsWithoutPayPalInSalesChannel(): void
     {
         $event = $this->createCmsPageLoadedEvent();
@@ -313,6 +338,17 @@ class ExpressCheckoutSubscriberTest extends TestCase
             $this->getExpectedExpressCheckoutButtonDataForAddProductEvents(),
             $actualExpressCheckoutButtonData
         );
+    }
+
+    public function testAddExpressCheckoutDataToPageletWithInactivePaymentMethod(): void
+    {
+        $event = $this->createQuickviewPageletLoadedEvent(false);
+
+        $this->getExpressCheckoutSubscriber()->addExpressCheckoutDataToPagelet($event);
+
+        /** @var ExpressCheckoutButtonData|null $actualExpressCheckoutButtonData */
+        $actualExpressCheckoutButtonData = $event->getPagelet()->getExtension(ExpressCheckoutSubscriber::PAYPAL_EXPRESS_CHECKOUT_BUTTON_DATA_EXTENSION_ID);
+        static::assertNull($actualExpressCheckoutButtonData);
     }
 
     public function testAddExpressCheckoutDataToPageletWithoutPayPalInSalesChannel(): void
@@ -400,6 +436,8 @@ class ExpressCheckoutSubscriberTest extends TestCase
         $cartService = $this->getContainer()->get(CartService::class);
         /** @var RouterInterface $router */
         $router = $this->getContainer()->get('router');
+        /** @var PaymentMethodUtil $paymentMethodUtil */
+        $paymentMethodUtil = $this->getContainer()->get(PaymentMethodUtil::class);
 
         return new ExpressCheckoutSubscriber(
             new PayPalExpressCheckoutDataService(
@@ -408,14 +446,11 @@ class ExpressCheckoutSubscriberTest extends TestCase
                 $router
             ),
             new SettingsServiceMock($settings ?? null),
-            new PaymentMethodUtil(
-                new PaymentMethodRepoMock(),
-                new SalesChannelRepoMock()
-            )
+            $paymentMethodUtil
         );
     }
 
-    private function createSalesChannelContext(bool $withItemList = false): SalesChannelContext
+    private function createSalesChannelContext(bool $withItemList = false, bool $paymentMethodActive = true): SalesChannelContext
     {
         $taxId = $this->createTaxId(Context::createDefaultContext());
         /** @var SalesChannelContextFactory $salesChannelContextFactory */
@@ -468,14 +503,34 @@ class ExpressCheckoutSubscriberTest extends TestCase
             $cartService->add($cart, $lineItem, $salesChannelContext);
         }
 
-        $paymentMethod = new PaymentMethodEntity();
-        $paymentMethod->setId(PaymentMethodRepoMock::PAYPAL_PAYMENT_METHOD_ID);
+        /** @var EntityRepositoryInterface $paymentMethodRepo */
+        $paymentMethodRepo = $this->getContainer()->get('payment_method.repository');
+        /** @var PaymentMethodUtil $paymentMethodUtil */
+        $paymentMethodUtil = $this->getContainer()->get(PaymentMethodUtil::class);
+        $paymentMethodId = $paymentMethodUtil->getPayPalPaymentMethodId($salesChannelContext->getContext());
+        static::assertNotNull($paymentMethodId);
 
-        $salesChannelEntity = $salesChannelContext->getSalesChannel();
-        $salesChannelEntity->setPaymentMethods(new PaymentMethodCollection([
-            $paymentMethod,
-        ]));
-        $salesChannelEntity->setId(Defaults::SALES_CHANNEL);
+        $paymentMethodRepo->update([[
+            'id' => $paymentMethodId,
+            'active' => $paymentMethodActive,
+        ]], $salesChannelContext->getContext());
+
+        /** @var EntityRepositoryInterface $salesChannelRepo */
+        $salesChannelRepo = $this->getContainer()->get('sales_channel.repository');
+
+        $paymentMethodIds = \array_unique(\array_merge(
+            $salesChannelContext->getSalesChannel()->getPaymentMethodIds() ?? [],
+            [$paymentMethodId]
+        ));
+
+        $salesChannelRepo->update([[
+            'id' => Defaults::SALES_CHANNEL,
+            'paymentMethods' => \array_map(static function (string $id) {
+                return ['id' => $id];
+            }, $paymentMethodIds),
+        ]], $salesChannelContext->getContext());
+
+        $salesChannelContext->getSalesChannel()->setPaymentMethodIds($paymentMethodIds);
 
         return $salesChannelContext;
     }
@@ -498,7 +553,7 @@ class ExpressCheckoutSubscriberTest extends TestCase
         return $taxId;
     }
 
-    private function createCmsPageLoadedEvent(bool $hasCmsPage = true): CmsPageLoadedEvent
+    private function createCmsPageLoadedEvent(bool $hasCmsPage = true, bool $paymentMethodActive = true): CmsPageLoadedEvent
     {
         $cmsPages = [];
         if ($hasCmsPage) {
@@ -512,13 +567,13 @@ class ExpressCheckoutSubscriberTest extends TestCase
         return new CmsPageLoadedEvent(
             new Request(),
             $result,
-            $this->createSalesChannelContext(true)
+            $this->createSalesChannelContext(true, $paymentMethodActive)
         );
     }
 
-    private function createQuickviewPageletLoadedEvent(): QuickviewPageletLoadedEvent
+    private function createQuickviewPageletLoadedEvent(bool $paymentMethodActive = true): QuickviewPageletLoadedEvent
     {
-        $salesChannelContext = $this->createSalesChannelContext(true);
+        $salesChannelContext = $this->createSalesChannelContext(true, $paymentMethodActive);
 
         /** @var EntityRepositoryInterface $productRepo */
         $productRepo = $this->getContainer()->get('product.repository');